prompt-search/requirements.txt
# Core
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0

# LLM APIs
anthropic>=0.40.0
requests>=2.31

# Local models — install separately with CUDA support if needed:
# GPU:  CMAKE_ARGS="-DLLAMA_CUDA=on" pip install llama-cpp-python --force-reinstall
# CPU:  pip install llama-cpp-python
llama-cpp-python>=0.2.0

# Torch for GPU detection (optional, CPU-only environments can skip)
# pip install torch --index-url https://download.pytorch.org/whl/cu121
torch>=2.0

# Data
datasets>=2.14
pandas>=2.0
numpy>=1.24

# Testing
pytest>=7.4
pytest-cov>=4.1

# Notebooks
jupyter>=1.0
ipywidgets>=8.0
matplotlib>=3.7

# Utilities
tqdm>=4.66

# RAG
sentence-transformers>=2.7
faiss-cpu>=1.8
